rolling back abstraction chain
