+++ src/Schema/Attribute/Attribute.php
@@ -133,53 +133,4 @@ class Attribute extends Item implements AttributeSingle
 
         return $this;
     }
-
-    /**
-     * @return Attribute
-     */
-    public static function loadAttribute(
-        SchemaReader $schemaReader,
-        Schema $schema,
-        DOMElement $node
-    ) {
-        $attribute = new self($schema, $node->getAttribute('name'));
-        $attribute->setDoc(SchemaReader::getDocumentation($node));
-        $schemaReader->fillItem($attribute, $node);
-
-        if ($node->hasAttribute('nillable')) {
-            $attribute->setNil($node->getAttribute('nillable') == 'true');
-        }
-        if ($node->hasAttribute('form')) {
-            $attribute->setQualified($node->getAttribute('form') == 'qualified');
-        }
-        if ($node->hasAttribute('use')) {
-            $attribute->setUse($node->getAttribute('use'));
-        }
-
-        return $attribute;
-    }
-
-    /**
-     * @return AttributeItem
-     */
-    public static function getAttributeFromAttributeOrRef(
-        SchemaReader $schemaReader,
-        DOMElement $childNode,
-        Schema $schema,
-        DOMElement $node
-    ) {
-        if ($childNode->hasAttribute('ref')) {
-            /**
-             * @var AttributeItem
-             */
-            $attribute = $schemaReader->findSomething('findAttribute', $schema, $node, $childNode->getAttribute('ref'));
-        } else {
-            /**
-             * @var Attribute
-             */
-            $attribute = self::loadAttribute($schemaReader, $schema, $childNode);
-        }
-
-        return $attribute;
-    }
 }
