+++ src/Schema/Attribute/Group.php
@@ -57,69 +57,4 @@ class Group implements AttributeItem, AttributeContainer
     {
         return $this->schema;
     }
-
-    /**
-     * @param string $attr
-     */
-    public static function findSomethingLikeThis(
-        SchemaReader $useThis,
-        Schema $schema,
-        DOMElement $node,
-        DOMElement $childNode,
-        AttributeContainer $addToThis
-    ) {
-        /**
-         * @var AttributeItem
-         */
-        $attribute = $useThis->findSomething('findAttributeGroup', $schema, $node, $childNode->getAttribute('ref'));
-        $addToThis->addAttribute($attribute);
-    }
-
-    /**
-     * @return \Closure
-     */
-    public static function loadAttributeGroup(
-        SchemaReader $schemaReader,
-        Schema $schema,
-        DOMElement $node
-    ) {
-        $attGroup = new self($schema, $node->getAttribute('name'));
-        $attGroup->setDoc(SchemaReader::getDocumentation($node));
-        $schema->addAttributeGroup($attGroup);
-
-        return function () use ($schemaReader, $schema, $node, $attGroup) {
-            SchemaReader::againstDOMNodeList(
-                $node,
-                function (
-                    DOMElement $node,
-                    DOMElement $childNode
-                ) use (
-                    $schemaReader,
-                    $schema,
-                    $attGroup
-                ) {
-                    switch ($childNode->localName) {
-                        case 'attribute':
-                            $attribute = Attribute::getAttributeFromAttributeOrRef(
-                                $schemaReader,
-                                $childNode,
-                                $schema,
-                                $node
-                            );
-                            $attGroup->addAttribute($attribute);
-                            break;
-                        case 'attributeGroup':
-                            self::findSomethingLikeThis(
-                                $schemaReader,
-                                $schema,
-                                $node,
-                                $childNode,
-                                $attGroup
-                            );
-                            break;
-                    }
-                }
-            );
-        };
-    }
 }
