+++ src/Schema/Type/BaseComplexType.php
@@ -12,20 +12,4 @@ use GoetasWebservices\XML\XSDReader\Schema\Schema;
 abstract class BaseComplexType extends Type implements AttributeContainer
 {
     use AttributeContainerTrait;
-
-    public function addAttributeFromAttributeOrRef(
-        SchemaReader $reader,
-        DOMElement $childNode,
-        Schema $schema,
-        DOMElement $node
-    ) {
-        $attribute = Attribute::getAttributeFromAttributeOrRef(
-            $reader,
-            $childNode,
-            $schema,
-            $node
-        );
-
-        $this->addAttribute($attribute);
-    }
 }
